Refactorisation packages et classes securité
